converts stories to fixtures and examples, adds axe violation tests
